+++ Enums/ExecutionMode.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Wifft 2023
+// Wifft licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace OCRStudio.Enums
+{
+    public enum ExecutionMode
+    {
+        Screenshoot = 0,
+        Recorder = 1
+    }
+}