+++ Views/OCROverlay/CaptureAreasPage.cs
@@ -1,0 +1,10 @@
+// Copyright (c) Wifft 2023
+// Wifft licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace OCRStudio.Views.OCROverlay
+{
+    public sealed partial class CaptureAreasPage : BasePage
+    {
+    }
+}